api/requirements.txt
fastapi
uvicorn
psycopg
hikari
python-dotenv